killed some MUI errors when porting HealthDashboard to ts
